*sigh*

+++ Sources/SwiftMsgPack/Decoder.swift
@@ -168,7 +168,7 @@ public extension Data {
         case 0xd4:
             let type = try stream.read8Bit()
             let data = try stream.read8Bit()
-            return [0 /*TODO(smolck)*/, type, data]
+            return [0 /*TODO(smolck)*/, type, UInt16(data)] // TODO(smolck)
         case 0xc7:
             let size = try stream.read8Bit()
             let type = try stream.read8Bit()
